requirements.txt
playwright
pandas
requests
python-dotenv
lxml
html5lib
BeautifulSoup4